+++ pages/dashboard/modules/index.js
@@ -7,4 +7,5 @@ export { PeekPanel } from "./peek-panel.js";
 export { MemoryPage } from "./memory-page.js";
 export { RecallPage } from "./recall-page.js";
 export { SystemPage } from "./system-page.js";
+export { PromptPage } from "./prompt-page.js";
 export * from "./utils.js";